Update and rename cv.html to index.html
I made this modification because when I clicked the GitHub Pages link, I encountered a 404 "File not found" error. This happened because GitHub Pages looks for a file named index.html in the root directory to render the website. Since my file was originally named cv.html, GitHub couldn’t locate it, resulting in the error.